backend/requirements.txt
anndata
antropy
arviz
bambi
bioinfokit
biopython
bokeh
cartopy
catboost
cellpose
cmdstanpy
csbdeep
dash
earthpy
elephant
fastapi
fastparquet
folium
geopandas
hdbscan
holoviews
igraph
imageio
ipywidgets
jax
joblib
json5
jupyterlab
leidenalg
lightgbm
loompy
matplotlib
mne
mudata
napari
neo
networkx
nibabel
nilearn
numpy
openpyxl
optuna
pandas
panel
patsy
pillow
pingouin
plotly
plotly-express
plotnine
pyproj
pyqtgraph
PyQt6
pyro-ppl
pytest
python-multipart
PyWavelets
pyyaml
rasterio
scanpy
scikit-image
scikit-learn
scipy
seaborn
shapely
simpy
stardist
statsmodels
streamlit
sympy
tqdm
tsfresh
umap-learn
uvicorn
xarray
xgboost
